+++ ui/port/src/main/java/androidx/ui/gestures/events/PointerEvent.kt
@@ -73,34 +73,34 @@ abstract class PointerEvent(
     // /// with no discernible pressure, to 1.0, indicating a touch with "normal"
     // /// pressure, and possibly beyond, indicating a stronger touch. For devices
     // /// that do not detect pressure (e.g. mice), returns 1.0.
-    val pressure: Double = 1.0,
+    val pressure: Float = 1.0f,
     // /// The minimum value that [pressure] can return for this pointer. For devices
     // /// that do not detect pressure (e.g. mice), returns 1.0. This will always be
     // /// a number less than or equal to 1.0.
-    val pressureMin: Double = 1.0,
+    val pressureMin: Float = 1.0f,
     // /// The maximum value that [pressure] can return for this pointer. For devices
     // /// that do not detect pressure (e.g. mice), returns 1.0. This will always be
     // /// a greater than or equal to 1.0.
-    val pressureMax: Double = 1.0,
+    val pressureMax: Float = 1.0f,
     // /// The distance of the detected object from the input surface (e.g. the
     // /// distance of a stylus or finger from a touch screen), in arbitrary units on
     // /// an arbitrary (not necessarily linear) scale. If the pointer is down, this
     // /// is 0.0 by definition.
-    val distance: Double = 0.0,
+    val distance: Float = 0.0f,
     // /// The maximum value that a distance can return for this pointer. If this
     // /// input device cannot detect "hover touch" input events, then this will be
     // /// 0.0.
-    val distanceMax: Double = 0.0,
+    val distanceMax: Float = 0.0f,
     // /// The radius of the contact ellipse along the major axis, in logical pixels.
-    val radiusMajor: Double = 0.0,
+    val radiusMajor: Float = 0.0f,
     // /// The radius of the contact ellipse along the minor axis, in logical pixels.
-    val radiusMinor: Double = 0.0,
+    val radiusMinor: Float = 0.0f,
     // /// The minimum value that could be reported for radiusMajor and radiusMinor
     // /// for this pointer, in logical pixels.
-    val radiusMin: Double = 0.0,
+    val radiusMin: Float = 0.0f,
     // /// The minimum value that could be reported for radiusMajor and radiusMinor
     // /// for this pointer, in logical pixels.
-    val radiusMax: Double = 0.0,
+    val radiusMax: Float = 0.0f,
     // /// For PointerDeviceKind.touch events:
     // ///
     // /// The angle of the contact ellipse, in radius in the range:
@@ -126,7 +126,7 @@ abstract class PointerEvent(
     // /// indicate that the stylus would go down in the negative y-axis direction;
     // /// pi/4 would indicate that the stylus goes up and to the right, -pi/2 would
     // /// indicate that the stylus goes to the left, etc).
-    val orientation: Double = 0.0,
+    val orientation: Float = 0.0f,
     // /// For PointerDeviceKind.stylus and PointerDeviceKind.invertedStylus events:
     // ///
     // /// The angle of the stylus, in radians in the range:
@@ -137,7 +137,7 @@ abstract class PointerEvent(
     // /// perpendicular to the input surface (thus 0.0 indicates the stylus is
     // /// orthogonal to the plane of the input surface, while pi/2 indicates that
     // /// the stylus is flat on that surface).
-    val tilt: Double = 0.0,
+    val tilt: Float = 0.0f,
     // /// We occasionally synthesize PointerEvents that aren"t exact translations
     // /// of [ui.PointerData] from the engine to cover small cross-OS discrepancies
     // /// in pointer behaviors.
@@ -151,7 +151,7 @@ abstract class PointerEvent(
     val synthesized: Boolean = false
 ) {
     // /// The minimum value that a distance can return for this pointer (always 0.0).
-    val distanceMin: Double = 0.0
+    val distanceMin: Float = 0.0f
 
     override fun toString(): String = "${this.javaClass.canonicalName}($position)"
